Spatial Queries

SpatialIndexing.SpatialQueryIterator — Type

Base abstract class for implementing spatial queries in N-dimensional space.

SpatialIndexing.QueryKind — Type

Specifies the kind of spatial data query.

SpatialIndexing.QueryMatch — Type

Specifies the result of spatial data query.

Base.findfirst — Function

findfirst(tree::RTree{T,N}, reg::Region{T,N}, [id]) where {T,N}

Find the element in the tree by its region (reg) and, optionally, its id. The region (MBR) of the element and reg should match exactly.

Returns the tuple of Leaf and position of the element or nothing.

Base.isempty — Function

isempty(tree::RTree, region::Region)

Check if there are tree elements inside region.

SpatialIndexing.contained_in — Function

contained_in(index::SpatialIndex, region::Region)

Get iterator for index elements contained in region.

SpatialIndexing.intersects_with — Function

intersects_with(index::SpatialIndex, region::Region)

Get iterator for index elements intersecting with region.