Spatial Primitives

SpatialIndexing.Region — Type

Base abstract class for implementing regions in N-dimensional space with dimensions of type T.

SpatialIndexing.Point — Type

N-dimensional point.

SpatialIndexing.Rect — Type

Rectangular region constrained by low[i]...high[i] in each of N dimensions.

SpatialIndexing.empty — Function

empty(::Type{T}) where T<:Region

Generate empty (uninitialized) region of type T.

SpatialIndexing.isvalid — Function

isvalid(a::Region)

Check that the parameters of a are valid and it defines a proper region.

SpatialIndexing.center — Function

center(a::Region)

The center point of a.

SpatialIndexing.area — Function

area(a::Region)

N-dimensional "area" (volume etc) of a.

SpatialIndexing.perimeter — Function

perimeter(a::Region)

The sum of the a sides.

SpatialIndexing.intersects — Function

intersects(a::Region, b::Region)

Check whether a intersects with b.

SpatialIndexing.contains — Function

contains(a::Region, b::Region)

Check whether a contains b.

SpatialIndexing.touches — Function

touches(a::Rect, b::Rect)

Check whether a and b rectangles touches from the inside (i.e. any low side touches low or high touches high).

Base.in — Function

in(a::Region, b::Region)

Check whether a is contained inside b.

SpatialIndexing.overlap_area — Function

overlap_area(a::Region, b::Region)

The area of MBR for a and b intersection.

SpatialIndexing.combined_area — Function

combined_area(a::Region, b::Region)

The area of MBR for the union of a and b.

SpatialIndexing.enlargement — Function

enlargement(a::Region, b::Region)

How much a grows when combined with b.

The difference between the MBR area of the a and b union and the MBR area of a.

SpatialIndexing.combine — Function

combine(a::Region, b::Region)

MBR that contains both a and b regions.

SpatialIndexing.intersect — Function

intersect(a::Region, b::Region)

MBR that is an intersection of a and b regions.