Simple Spatial Index

SimpleSpatialIndex stores all data elements in a vector. So, while insertion of new data takes constant time, the time of spatial searches grows linearly with the number of elements. This spatial index is intended as a reference implementation for benchmarking and not recommended for production usage.

SpatialIndexing.SimpleSpatialIndex — Type

Vector-based SpatialIndex. While insertion is O(1), the search is O(N).

Generally should not be used, except for performance comparisons or when the number of stored elements is expected to be very small (<100).